op-challenger: Split library code to a separate file.
