+++ op-challenger/game/keccak/matrix/testdata/commitments.json
